requirements.txt
asgiref==3.3.1
certifi==2021.5.30
chardet==3.0.4
charset-normalizer==2.0.4
Django==3.1.6
django-bootstrap-modal-forms==2.0.1
django-ckeditor==6.1.0
django-classy-tags==0.6.2
django-embed-video==1.4.0
django-environ==0.4.5
django-google-translate==1.1
django-js-asset==1.2.2
django-widget-tweaks==1.4.8
django-youtube-thumbnail==1.0.1
googletrans==3.0.0
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.12.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
jsonfield==3.1.0
numpy==1.21.2
pafy==0.3.82
pandas==1.3.3
Pillow==8.1.0
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytube==11.0.1
pytz==2021.1
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.1
urllib3==1.26.6
youtube-dl==2021.6.6
